aws: Refactor AWS client code
