+++ pkg/aws/client.go
@@ -27,7 +27,6 @@ import (
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/aws/client"
 	"github.com/aws/aws-sdk-go/aws/credentials"
-	"github.com/aws/aws-sdk-go/aws/request"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/cloudformation"
 	"github.com/aws/aws-sdk-go/service/cloudformation/cloudformationiface"
@@ -334,165 +333,6 @@ func (c *awsClient) ValidateCredentials() (bool, bool, error) {
 	return true, false, nil
 }
 
-// Ensure osdCcsAdmin IAM user is created
-func (c *awsClient) EnsureOsdCcsAdminUser(stackName string, adminUserName string, awsRegion string) (bool, error) {
-	userExists := true
-	regionForInit, err := c.GetClusterRegionTagForUser(adminUserName)
-	if err != nil {
-		//If user doesnt exists proceed normally
-		//If users exists and tag is not present then add the tag
-		//if user exists and tag is present then proceed normally
-		if aerr, ok := err.(awserr.Error); ok {
-			switch aerr.Code() {
-			case iam.ErrCodeNoSuchEntityException:
-				userExists = false
-			default:
-				return false, err
-			}
-		}
-	}
-	if userExists {
-		if regionForInit == "" {
-			err = c.TagUserRegion(adminUserName, DefaultRegion)
-			if err != nil {
-				return false, err
-			}
-		}
-		return false, nil
-	}
-	// Check already existing cloudformation stack status
-	stackReady, stackStatus, err := c.CheckStackReadyOrNotExisting(stackName)
-	if err != nil {
-		return false, err
-	}
-
-	// Read cloudformation template
-	cfTemplateBody, err := readCFTemplate()
-	if err != nil {
-		return false, err
-	}
-
-	// If stack CREATE_COMPLETE or UPGRADE_COMPLETE the stack is already create
-	// try to update it in case the cloudformation template has changed
-	if stackStatus != nil {
-		if (*stackStatus == cloudformation.StackStatusCreateComplete) ||
-			(*stackStatus == cloudformation.StackStatusUpdateComplete) {
-			_, err = c.UpdateStack(cfTemplateBody, stackName)
-			if err != nil {
-				return false, err
-			}
-
-			return false, nil
-		}
-	}
-
-	// If the Cloudformation stack isn't ready, make sure the IAM user
-	// doesn't exist or the Cloudformation stack create will fail
-	if !stackReady {
-		err = c.CheckAdminUserNotExisting(adminUserName)
-		if err != nil {
-			return false, err
-		}
-	}
-
-	// Create stack
-	_, err = c.CreateStack(cfTemplateBody, stackName)
-	if err != nil {
-		return false, err
-	}
-
-	err = c.TagUserRegion(adminUserName, awsRegion)
-	if err != nil {
-		return false, err
-	}
-	return true, nil
-}
-
-func (c *awsClient) CreateStack(cfTemplateBody, stackName string) (bool, error) {
-	// Create cloudformation stack
-	_, err := c.cfClient.CreateStack(buildCreateStackInput(cfTemplateBody, stackName))
-	if err != nil {
-		return false, err
-	}
-
-	// Wait until cloudformation stack creates
-	err = c.cfClient.WaitUntilStackCreateComplete(&cloudformation.DescribeStacksInput{
-		StackName: aws.String(stackName),
-	})
-	if err != nil {
-		switch typed := err.(type) {
-		case awserr.Error:
-			// Waiter reached maximum attempts waiting for the resource to be ready
-			if typed.Code() == request.WaiterResourceNotReadyErrorCode {
-				c.logger.Errorf("Max retries reached waiting for stack to create")
-				return false, err
-			}
-		}
-		return false, err
-	}
-
-	return true, nil
-}
-
-func (c *awsClient) UpdateStack(cfTemplateBody, stackName string) (bool, error) {
-	_, err := c.cfClient.UpdateStack(buildUpdateStackInput(cfTemplateBody, stackName))
-	if err != nil {
-		switch typed := err.(type) {
-		case awserr.Error:
-			// Exit true if there is no update to be performed on the cloudformation stack
-			if typed.Code() == "ValidationError" {
-				if typed.Message() == "No updates are to be performed." {
-					return true, nil
-				}
-			}
-		}
-		return false, err
-	}
-
-	// Wait for CloudFormation update to complete
-	err = c.cfClient.WaitUntilStackUpdateComplete(&cloudformation.DescribeStacksInput{
-		StackName: aws.String(stackName),
-	})
-
-	if err != nil {
-		switch typed := err.(type) {
-		case awserr.Error:
-			// Waiter reached maximum attempts waiting for the resource to be ready
-			if typed.Code() == request.WaiterResourceNotReadyErrorCode {
-				c.logger.Errorf("Max retries reached waiting for stack to create")
-				return false, err
-			}
-		}
-		return false, err
-	}
-
-	return true, err
-}
-
-func (c *awsClient) CheckStackReadyOrNotExisting(stackName string) (stackReady bool, status *string, err error) {
-	stackList, err := c.cfClient.ListStacks(&cloudformation.ListStacksInput{})
-	if err != nil {
-		return false, nil, err
-	}
-
-	for _, summary := range stackList.StackSummaries {
-		if *summary.StackName == stackName {
-			if (*summary.StackStatus == cloudformation.StackStatusCreateComplete) ||
-				(*summary.StackStatus == cloudformation.StackStatusUpdateComplete) {
-				return true, summary.StackStatus, nil
-			}
-			if *summary.StackStatus != cloudformation.StackStatusDeleteComplete {
-				return false, summary.StackStatus, fmt.Errorf("Error creating user: Cloudformation stack %s exists "+
-					"with status %s. Expected status is %s.\n"+
-					"Ensure %s CloudFormation Stack does not exist, then retry with\n"+
-					"rosa init --delete-stack; rosa init",
-					*summary.StackName, *summary.StackStatus, cloudformation.StackStatusCreateComplete, *summary.StackName)
-			}
-		}
-	}
-	return false, nil, nil
-}
-
 func (c *awsClient) CheckAdminUserNotExisting(userName string) (err error) {
 	userList, err := c.iamClient.ListUsers(&iam.ListUsersInput{})
 	if err != nil {
@@ -514,42 +354,6 @@ func (c *awsClient) CheckAdminUserExists(userName string) (err error) {
 	if err != nil {
 		return err
 	}
-	return nil
-}
-
-func (c *awsClient) DeleteOsdCcsAdminUser(stackName string) error {
-	deleteStackInput := &cloudformation.DeleteStackInput{
-		StackName: aws.String(stackName),
-	}
-
-	// Delete cloudformation stack
-	_, err := c.cfClient.DeleteStack(deleteStackInput)
-	if err != nil {
-		switch typed := err.(type) {
-		case awserr.Error:
-			if typed.Code() == cloudformation.ErrCodeTokenAlreadyExistsException {
-				return nil
-			}
-		}
-		return err
-	}
-
-	// Wait until cloudformation stack deletes
-	err = c.cfClient.WaitUntilStackDeleteComplete(&cloudformation.DescribeStacksInput{
-		StackName: aws.String(stackName),
-	})
-	if err != nil {
-		switch typed := err.(type) {
-		case awserr.Error:
-			// Waiter reached maximum attempts waiting for the resource to be ready
-			if typed.Code() == request.WaiterResourceNotReadyErrorCode {
-				c.logger.Errorf("Max retries reached waiting for stack to delete")
-				return err
-			}
-		}
-		return err
-	}
-
 	return nil
 }
 
@@ -733,94 +537,4 @@ func (c *awsClient) DeleteAccessKeys(username string) error {
 
 	// Complete, deleted all accesskeys for `username`
 	return nil
-}
-
-// ValidateQuota
-func (c *awsClient) ValidateQuota() (bool, error) {
-	for _, quota := range serviceQuotaServices {
-		serviceQuotas, err := ListServiceQuotas(c, quota.ServiceCode)
-		if err != nil {
-			return false, fmt.Errorf("Error listing AWS service quotas: %s %v", quota.ServiceCode, err)
-		}
-
-		serviceQuota, err := GetServiceQuota(serviceQuotas, quota.QuotaCode)
-		if err != nil || serviceQuota == nil || (*serviceQuota).Value == nil {
-			return false, fmt.Errorf("Error getting AWS service quota: %s %v", quota.ServiceCode, err)
-		}
-
-		if *serviceQuota.Value < *quota.DesiredValue {
-			return false, fmt.Errorf(
-				"Service %s quota code %s %s not valid, expected quota of at least %d, but got %d",
-				quota.ServiceCode, quota.QuotaCode, quota.QuotaName,
-				int(*quota.DesiredValue), int(*serviceQuota.Value))
-		}
-
-		c.logger.Debug(fmt.Sprintf("Service %s quota code %s is ok", quota.ServiceCode, quota.QuotaCode))
-	}
-
-	return true, nil
-}
-
-// ValidateSCP attempts to validate SCP policies by ensuring we have the correct permissions
-func (c *awsClient) ValidateSCP(target *string) (bool, error) {
-	scpPolicyPath := "templates/policies/osd_scp_policy.json"
-
-	sParams := &SimulateParams{
-		Region: *c.awsSession.Config.Region,
-	}
-
-	// Read installer permissions and OSD SCP Policy permissions
-	osdPolicyDocument := readSCPPolicy(scpPolicyPath)
-	policyDocuments := []PolicyDocument{osdPolicyDocument}
-
-	// Get Creator details
-	creator, err := c.GetCreator()
-	if err != nil {
-		return false, err
-	}
-
-	// Find target user
-	var targetUserARN arn.ARN
-	if target == nil {
-		var err error
-		callerIdentity, _, err := getClientDetails(c)
-		if err != nil {
-			return false, fmt.Errorf("getClientDetails: %v\n"+
-				"Run 'rosa init' and try again", err)
-		}
-		targetUserARN, err = arn.Parse(*callerIdentity.Arn)
-		if err != nil {
-			return false, fmt.Errorf("unable to parse caller ARN %v", err)
-		}
-		// If the client is using STS credentials want to validate the role
-		// the user has assumed. GetCreator() resolves that for us and updates
-		// the ARN
-		if creator.IsSTS {
-			targetUserARN, err = arn.Parse(creator.ARN)
-			if err != nil {
-				return false, err
-			}
-		}
-	} else {
-		targetIAMOutput, err := c.iamClient.GetUser(&iam.GetUserInput{UserName: target})
-		if err != nil {
-			return false, fmt.Errorf("iamClient.GetUser: %v\n"+
-				"To reset the '%s' account, run 'rosa init --delete-stack' and try again", *target, err)
-		}
-		targetUserARN, err = arn.Parse(*targetIAMOutput.User.Arn)
-		if err != nil {
-			return false, fmt.Errorf("unable to parse caller ARN %v", err)
-		}
-	}
-
-	// Validate permissions
-	hasPermissions, err := validatePolicyDocuments(c, targetUserARN.String(), policyDocuments, sParams)
-	if err != nil {
-		return false, err
-	}
-	if !hasPermissions {
-		return false, err
-	}
-
-	return true, nil
 }
